requirements.txt
Pillow==7.1.2
PyQt5==5.15.0
PyQt5-sip==12.8.0